refactored code into components & updated code structure
