+++ src/reducers/postReducer.jsx
@@ -1,0 +1,10 @@
+export const postReducer = (postState,action) =>{
+    switch(action.type){
+        case 'SET_ALL_POSTS': return {...postState, posts:action.payload}
+        case 'ADD_NEW_POST': return {...postState, posts:action.payload}
+        case 'GET_ALL_BOOKMARKS': return {...postState, bookmarks:action.payload}
+        case 'EDIT_POST': return {...postState,posts:action.payload}
+        case 'DELETE_POST': return {...postState,posts:action.payload}
+        default: return postState;
+    }
+}
